+++ src/azure-cli/azure/cli/command_modules/reservations/_client_factory.py
@@ -6,7 +6,7 @@
 
 def cf_reservations(cli_ctx, **_):
     from azure.cli.core.commands.client_factory import get_mgmt_service_client
-    from azure.mgmt.reservations.azure_reservation_api import AzureReservationAPI
+    from azure.mgmt.reservations import AzureReservationAPI
     return get_mgmt_service_client(cli_ctx, AzureReservationAPI, subscription_bound=False)